+++ worker/packages/logging/package.js
@@ -1,9 +1,0 @@
-Package.describe({
-  summary: "Shared worker logging facility",
-  internal: true
-});
-
-Package.on_use(function (api) {
-  api.use('coffeescript');
-  api.add_files('logging.coffee', 'client');
-});