+++ History.md
@@ -1,3 +1,11 @@
+## v0.0.4
+
+Fix problem with confusing Meteorite.  Move common code used by both
+the offline data client and the shared web worker agent into
+offline-common, and the app to build the shared web worker into
+offline-worker.
+
+
 ## v0.0.3
 
 Use a shared web worker when supported for the agent.  #17